rename(list-detail): 폴더 구조 개선
