+++ src/app/(route)/list/[id]/_components/LIST_DETAIL_CONST.ts
