+++ src/app/(route)/list/[id]/_components/PostActionMenu/PostActionMenu.tsx
@@ -6,11 +6,11 @@ import { cn } from "@/utils";
 import { Button, Icon } from "@/components/common";
 import { useDeleteDetailPost, usePutPostStatus } from "@/api/fetch/post";
 import ModalLayout from "@/components/common/Modal/_internal/ModalLayout";
-import { ACTION_MENU } from "./ACTION_MENU_STYLES";
 import { PostReportBlockActions, ReportModal } from "@/components/domain";
 import { useGetUsersMe } from "@/api/fetch/user";
 import { PostActionData } from "../../_types/PostActionType";
 import UserBlockModal from "@/components/domain/PostReportBlockActions/UserBlockModal/UserBlockModal";
+import { ACTION_MENU } from "../LIST_DETAIL_CONST";
 
 interface PostOptionBoxProps {
   open: boolean;
