--- src/app/(route)/list/[id]/_components/PostDetail/_internal/index.ts
@@ -1,2 +1,0 @@
-export { default as PostDetailBody } from "./PostDetailBody/PostDetailBody";
-export { default as PostDetailPreviewKakaoMap } from "./PostDetailPreviewKakaoMap/PostDetailPreviewKakaoMap";
