+++ src/app/(route)/list/[id]/_components/_internal/ImageSection/ImageSection.tsx
