+++ src/app/(route)/list/[id]/_components/_internal/ImageSection/swiper-pagination.css
