+++ src/app/(route)/list/[id]/_components/_internal/PostChipSection/PostChipSection.tsx
