+++ src/app/(route)/list/[id]/_components/_internal/PostDetailPreviewKakaoMap/PostDetailPreviewKakaoMap.tsx
