+++ src/app/(route)/list/[id]/_components/_internal/index.ts
@@ -1,3 +1,6 @@
 export { default as DetailSkeleton } from "./DetailSkeleton/DetailSkeleton";
 export { default as ErrorSimilarSection } from "./ErrorSimilarSection/ErrorSimilarSection";
 export { default as SimilarItemsList } from "./SimilarItemsList/SimilarItemsList";
+export { default as PostDetailBody } from "./PostDetailBody/PostDetailBody";
+export { default as PostDetailPreviewKakaoMap } from "./PostDetailPreviewKakaoMap/PostDetailPreviewKakaoMap";
+export { default as ImageSection } from "./ImageSection/ImageSection";
