+++ src/app/(route)/list/_components/DefaultListSearch/DefaultListSearch.tsx
@@ -7,8 +7,8 @@ import { useGetSearchKeyword } from "@/api/fetch/post";
 import { ErrorBoundary } from "@/app/ErrorBoundary";
 import { InputSearch } from "@/components/common";
 import { useInfiniteScroll } from "@/hooks";
-import { PostSearchView } from "./_internal";
 import { trackSearch } from "@/utils/analytics/analytics";
+import { PostSearchView } from "../_internal";
 
 const DefaultListSearch = () => {
   const router = useRouter();
