--- src/app/(route)/list/_components/DefaultListSearch/_internal/index.ts
@@ -1,1 +1,0 @@
-export { default as PostSearchView } from "./PostSearchView/PostSearchView";
