+++ src/app/(route)/list/_components/_internal/PostSearchEmpty/PostSearchEmpty.tsx
