+++ src/app/(route)/list/_components/_internal/PostSearchView/PostSearchView.tsx
