+++ src/app/(route)/list/_components/_internal/index.ts
@@ -1,1 +1,3 @@
 export { default as DefaultList } from "./DefaultList/DefaultList";
+export { default as PostSearchEmpty } from "./PostSearchEmpty/PostSearchEmpty";
+export { default as PostSearchView } from "./PostSearchView/PostSearchView";
